requirements.txt
streamlit
python-dotenv
textblob
langchain-google-genai
